Change the duration after some research
Under  the  Limitation  Act  (Cap.  163),  actions  founded  on  a  contract  (amongst others)  must  be  brought  within  6  years  from  the  date  on  which  the  cause  of action accrued.  Hence an organisation may wish to retain records relating to its contracts for  7  years  from  the  date  of termination of  the  contract  and  possibly for  a  longer  period  if  an  investigation  or  legal  proceedings  should  commence within that period.